pyproject.toml
[tool.poetry]
name = "rentalhelper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.4"
djangorestframework = "^3.11.0"
markdown = "^3.2.1"
django-filter = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
